refactor: 코드 컨벤션에 맞도록 정리, 폴더 구조 정리
