+++ src/app/groups/[id]/management/_components/ManagementBtns.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { ModalModeType } from './ManagementContents';
 
 interface ManagementBtnsProps {
