+++ src/app/groups/[id]/management/_components/modal/LeaveGroupModal.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const LeaveGroupModal = () => {
   return (
