+++ src/app/groups/[id]/management/members/_components/CurMemberList.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import MemberCard from './MemberCard';
 
 const CurMemberList = () => {
